+++ src/simmate/apps/cas_registry/components/__init__.py
@@ -1,3 +1,3 @@
 # -*- coding: utf-8 -*-
 
-from .cas_registry_molecule_table import CasRegistryMoleculeTable
+from .molecule import CasRegistryMoleculeComponent