+++ src/simmate/apps/emolecules/components/__init__.py
@@ -1,4 +1,4 @@
 # -*- coding: utf-8 -*-
 
-from .emolecules_molecule_table import EmoleculesMoleculeTable
-from .emolecules_supplier_offer_table import EmoleculesSupplierOfferTable
+from .molecule import EmoleculesMoleculeComponent
+from .supplier_offer import EmoleculesSupplierOfferComponent