+++ src/simmate/apps/evolution/components/__init__.py
@@ -1,8 +1,6 @@
 # -*- coding: utf-8 -*-
 
-from .chemical_system_search_table import ChemicalSystemSearchTable
-from .fixed_composition_search_table import FixedCompositionSearchTable
-from .steadystate_source_table import SteadystateSourceTable
-from .variable_nsites_composition_search_table import (
-    VariableNsitesCompositionSearchTable,
-)
+from .chemical_system_search import ChemicalSystemSearchComponent
+from .fixed_composition_search import FixedCompositionSearchComponent
+from .steadystate_source import SteadystateSourceComponent
+from .variable_nsites_composition_search import VariableNsitesCompositionSearchComponent