+++ src/simmate/apps/materials_project/components/__init__.py
@@ -1,3 +1,3 @@
 # -*- coding: utf-8 -*-
 
-from .matproj_structure_table import MatprojStructureTable
+from .structure import MatprojStructureComponent