+++ src/simmate/apps/oqmd/components/__init__.py
@@ -1,3 +1,3 @@
 # -*- coding: utf-8 -*-
 
-from .oqmd_structure_table import OqmdStructureTable
+from .structure import OqmdStructureComponent