+++ src/simmate/apps/price_catalog/components/__init__.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
 
-from .price_point_table import PricePointTable
-from .priced_item_table import PricedItemTable
-from .priced_items_report import PricedItemsReport
+from .price_point import PricePointComponent
+from .priced_item import PricedItemComponent
+from .priced_items_report import PricedItemsReportComponent