+++ src/simmate/website/htmx/components/dynamic_table_form_mixins/__init__.py
@@ -1,7 +1,0 @@
-# -*- coding: utf-8 -*-
-
-from .create import CreateMixin
-from .create_many import CreateManyMixin
-from .search import SearchMixin
-from .update import UpdateMixin
-from .update_many import UpdateManyMixin