+++ src/simmate/website/workflow_explorer/components/__init__.py
@@ -1,4 +1,17 @@
 # -*- coding: utf-8 -*-
 
-from .workflow_search_form import WorkflowSearchForm
-from .workflow_submission_form import WorkflowSubmissionForm
+from .results import (
+    BandStructureCalcComponent,
+    DensityofStatesCalcComponent,
+    DiffusionAnalysisComponent,
+    DynamicsComponent,
+    DynamicsIonicStepComponent,
+    IonicStepComponent,
+    MigrationHopComponent,
+    MigrationImageComponent,
+    PopulationAnalysisComponent,
+    RelaxationComponent,
+    StaticEnergyComponent,
+)
+from .search import WorkflowSearchComponent
+from .submission import WorkflowSubmissionComponent